backend/requirements.txt
fastapi
uvicorn
chromadb
langchain
langgraph
langchain-google-genai
langchain-groq
python-multipart
python-dotenv
psycopg2-binary
pydantic
requests
mlflow
scikit-learn
pytest
httpx
Pillow
numpy
pandas
